converts mohammad's FlagDetection to new class, in new file and moves test to misc.
